work around new behavior of python-gconf
gconf's get_string used to return "" on undefined objects and now
returns None. there was no visible impact on functionality, but
exceptions flew out of the main loop when the metacity window was
created.

no measures were taken to find out exactly when the change was
introduced, as the fix is easy enough to just be applied.