minimal changes from cglita
this patch takes just the core of the cglita branch modifications, feeds
the newly gathered mode name information into a NamedSize (to replace
the old Size objects). thus, with minimal code changes, the name-based
mode selection arguments can be passed around.

this drops all the detailed mode selection, including the gui changes
that came with it. (the frequencies displayed were pixel clocks anyway,
which are hardly useful to anyone).